Move repository unblock to ESIntegTestCase cluster cleanup
